+++ packages/app/lab/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.126](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.125...@opendaw/lab@0.0.126) (2026-03-12)
+
+**Note:** Version bump only for package @opendaw/lab
+
+## [0.0.125](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.124...@opendaw/lab@0.0.125) (2026-03-12)
+
+**Note:** Version bump only for package @opendaw/lab
+
+## [0.0.124](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.123...@opendaw/lab@0.0.124) (2026-03-11)
+
+**Note:** Version bump only for package @opendaw/lab
+
+## [0.0.123](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.122...@opendaw/lab@0.0.123) (2026-03-11)
+
+**Note:** Version bump only for package @opendaw/lab
+
+## [0.0.122](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.121...@opendaw/lab@0.0.122) (2026-03-10)
+
+**Note:** Version bump only for package @opendaw/lab
+
 ## [0.0.121](https://github.com/andremichelle/openDAW/compare/@opendaw/lab@0.0.120...@opendaw/lab@0.0.121) (2026-03-07)
 
 **Note:** Version bump only for package @opendaw/lab